Initial writeup pre-code
Ok guys Heres the initial outline, if someone wants to hop on the regression coding i'd appreciate it, I can help out as well. I've outlined the process of least-squares but there's some online resources as well if needed. We need to run regression on the data set attached for each variable noted in the writeup, theres 9 total variables (or the last 9 columns of the attached excel file) that need to be compared against each other both as y and x (72 runs total). If this makes no sense I probably did a horrible job explaining and I'm happy to clarify or meet up and chat. Please let me know if anything looks inaccurate in the writeup as well.